archimed-back/requirements.txt
Django==3.2.25
djangorestframework==3.15.0
django-cors-headers==4.1.0